+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * fix for searching in unpacked package a binary file to partial match is used instead of full match
 * add a disable version checking checkbox
 * modify a task name to clickable for opening a GitHub repository in a task list
+* Change a drawer type to dismissible